Copy files instead of moving them
Some people may loose the files they keep in a specific folder moving their jars to the pack. It would maybe be a good idea to copy the files to the pack instead of moving them.
Not sure if my syntax is correct, but I try my best (being beginner in coding).
What do you think?